requirements.txt


gunicorn
django
djangorestframework
whitenoise
dj-database-url
ipython
